requirements.txt
beautifulsoup4==4.4.1
python-dateutil==2.5.0
requests==2.9.1
tinydb==3.1.3
Flask==0.10.1